+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/groq
 
+## 3.0.0-beta.21
+
+### Patch Changes
+
+- Updated dependencies [293a6b7]
+  - @ai-sdk/provider-utils@4.0.0-beta.21
+
 ## 3.0.0-beta.20
 
 ### Patch Changes