+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,34 @@
 
 ### Patch Changes
 
+- Updated dependencies [293a6b7]
+  - ai@6.0.0-beta.74
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [754df61]
+  - ai@6.0.0-beta.73
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [eca63f3]
+  - ai@6.0.0-beta.72
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [077aea3]
+  - ai@6.0.0-beta.71
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [d7bae86]
   - ai@6.0.0-beta.70